Fix(characters): fix orb controller partially
If you enable "Orb controllers" filter, and pick any of the orbs, then
deselect that orb, the filtered units will be less than before, even if
no orbs are selected. This is due to the algorithm for the orbs using a
different regex than what is used by the main filter. This fixes this
issue, but does not fix all the bugs with the orb controller (which
would require rewording all orb controllers to follow a strict format).